Third commit - Changes to Navigational bar and CSS
Navigation bar changes - created nested lists so that the containers for the navigation drop downs are connected to the navigation bar irrespective of screen size.
Originally had the hero/main section in a grid structure but have changed to flexbox. Further additions required for desktop.
Mobile media queries also required in further implementation.